+++ sigma/modules/searches/dictionary/urbandictionary.py
@@ -1,5 +1,5 @@
 ﻿# Apex Sigma: The Database Giant Discord Bot.
-# Copyright (C) 2018  Lucia's Cipher
+# Copyright (C) 2019  Lucia's Cipher
 #
 # This program is free software: you can redistribute it and/or modify
 # it under the terms of the GNU General Public License as published by
@@ -25,12 +25,11 @@ from sigma.core.utilities.generic_responses import error, not_found
 
 
 async def urbandictionary(cmd: SigmaCommand, pld: CommandPayload):
-    api_key = cmd.cfg.get('api_key')
-    if api_key:
+    if cmd.cfg.api_key:
         if pld.args:
             ud_input = ' '.join(pld.args).lower()
             url = "https://mashape-community-urban-dictionary.p.mashape.com/define?term=" + ud_input
-            headers = {'X-Mashape-Key': api_key, 'Accept': 'text/plain'}
+            headers = {'X-Mashape-Key': cmd.cfg.api_key, 'Accept': 'text/plain'}
             async with aiohttp.ClientSession() as session:
                 async with session.get(url, headers=headers) as data_response:
                     data = await data_response.read()