+++ sigma/modules/searches/food/foodrecipe.py
@@ -1,5 +1,5 @@
 ﻿# Apex Sigma: The Database Giant Discord Bot.
-# Copyright (C) 2018  Lucia's Cipher
+# Copyright (C) 2019  Lucia's Cipher
 #
 # This program is free software: you can redistribute it and/or modify
 # it under the terms of the GNU General Public License as published by
@@ -25,11 +25,10 @@ from sigma.core.utilities.generic_responses import error, not_found
 
 
 async def foodrecipe(cmd: SigmaCommand, pld: CommandPayload):
-    if 'api_key' in cmd.cfg:
-        api_key = cmd.cfg['api_key']
+    if cmd.cfg.api_key:
         if pld.args:
             search = ' '.join(pld.args)
-            url = f'http://food2fork.com/api/search?key={api_key}&q={search}'
+            url = f'http://food2fork.com/api/search?key={cmd.cfg.api_key}&q={search}'
             async with aiohttp.ClientSession() as session:
                 async with session.get(url) as data:
                     search_data = await data.read()