+++ sigma/modules/utilities/mathematics/wolframalpha.py
@@ -1,5 +1,5 @@
 ﻿# Apex Sigma: The Database Giant Discord Bot.
-# Copyright (C) 2018  Lucia's Cipher
+# Copyright (C) 2019  Lucia's Cipher
 #
 # This program is free software: you can redistribute it and/or modify
 # it under the terms of the GNU General Public License as published by
@@ -72,13 +72,12 @@ async def send_response(message: discord.Message, init: discord.Message or None,
 
 
 async def wolframalpha(cmd: SigmaCommand, pld: CommandPayload):
-    api_key = cmd.cfg.get('app_id')
     init_message = None
-    if api_key:
+    if cmd.cfg.app_id:
         if pld.msg.channel.id not in math_chs:
             if pld.args:
                 query = make_safe_query(pld.args)
-                url = f'{api_url}{query}&appid={api_key}'
+                url = f'{api_url}{query}&appid={cmd.cfg.app_id}'
                 init_response = discord.Embed(color=0xff7e00)
                 init_response.set_author(name='Processing request...', icon_url=wolfram_icon)
                 init_message = await pld.msg.channel.send(embed=init_response)